webapi/webapi-messaging-sysapps-tests/messaging/MessagingCursor.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu.Jiazhen <jiazhenx.shentu@intel.com>
        Jianfeng.Xu <jianfengx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>Messaging Test: MessagingCursor</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/messaging/#messagingcursor-interface">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/support.js"></script>
<div id="log"></div>
<script>

var filter = {
  type: "sms"
};
var options = {
  sortBy:"data",
  sortOrder:"descending",
  limit : 10
};

var t1 = async_test("Check if the element attribute of messagingcursor is valid", {timeout: 2000});
var t2 = async_test("Check if the next method of messagingcursor is valid", {timeout: 2000});
var t3 = async_test("Check if the previous method of messagingcursor is valid", {timeout: 2000});

t1.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, options).then(function (data) {
    t1.step(function () {
      assert_true(date.element.length > 0, "Therer is no message can be find");
      assert_true(!!data.element[0], "The element of MessagingCursor");
    });
    t1.done();
  }, function (error) {
    t1.step(function () {
      assert_unreached(error.message);
    });
    t1.done();
  });
});

t2.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, options).then(function (data) {
    t2.step(function () {
      assert_true(date.element.length > 0, "Therer is no message can be find");
      var beforeNext = data.element[0];
      data.next();
      var afterNext = data.element[0];
      assert_not_equals(beforeNext, afterNext, "After run next method, the element shoud be changed");
    });

    t2.done();
  }, function (error) {
    t2.step(function () {
      assert_unreached(error.message);
    });
    t2.done();
  });
});

t3.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, options).then(function (data) {
    t3.step(function () {
      assert_true(date.element.length > 0, "Therer is no message can be find");
      var beforePrevious = data.element[0];
      data.next();
      data.previous();
      var afterPrevious = data.element[0];
      assert_equals(beforePrevious, afterPrevious, "After run next method and previous method, the element shoud be same");
    });
    t3.done();
  }, function (error) {
    t3.step(function () {
      assert_unreached(error.message);
    });
    t3.done();
  });
});

</script>
